test: typing 도메인 관련 테스트를 수행합니다.
- 정확도 공식에 대한 테스트를 진행합니다.
- ROLE 오타 정정에 대한 테스트를 진행합니다.
- 정확도(acc)에 대한 자료형 변환에 대한 테스트를 진행합니다.

issue #51

+++ src/test/java/dasi/typing/api/service/ranking/RankingServiceTest.java
@@ -111,7 +111,7 @@ class RankingServiceTest {
   private Typing createTyping(int cpm, int acc, Member member, Phrase phrase) {
     return Typing.builder()
         .cpm(cpm)
-        .acc(acc)
+        .acc((double) acc)
         .wpm(0)
         .maxCpm(0)
         .member(member)
